Add CI linting and formatting (#9)
Add CI step that checks formatting and linting. 

Improved Makefile infrastructure to enable `make lint`, `make
check-format`, `make install`, and `make install-dev` in the root
directory which then lints/formats/installs all subprojects. Fixed lint
and format errors that revealed.

Added a script to upgrade dependencies project-wide.

Changed centrality and conclib to require python 3.10+ instead of 3.11+
(that is the version on the action runner and I decided 3.11 is a little
too modern to be the right choice).