Add preview deployment URL as PR comment (#13)
Deploying a preview instance now saves the preview url to
`/tmp/preview-url`. Added an extra step to preview deployment that adds
a PR comment with the preview URL. It will create a new comment if no
URL comment exists or will overwrite an existing comment if it does. It
checks whether a comment exists already by matching a unique string