Add complete VM lifecycle (incl GPU support) (#23)
Define a proper VM lifecycle, starting with a registration step that
sends the machine specs to the control plane, a limbo state where we are
not receiving heartbeats but the machine might reconnect, and a dead
state where the machine-id can be reused. Add a doc page with
information about the VM lifecycle.

Added support for GPU instances (rapidui still needs to be updated to
use this).

## Major changes
- Add rules around registration to prevent multiple different machines
from using the same machine id at the same time.
- Add ability to fake machine info, e.g. simulating having a GPU
- Automatically transition VMs from `limbo` to `dead` after some time
and delete them from the vm_info table
- Prevent heartbeats from occurring before registration
- VMAgent startup waits until the control plane is healthy
- Rename vm_heartbeat table to vm_info

## Minor changes
- Added test utils for common assertions to have information error
messages.
- Add support for optional types in CentralityConfig (optional subconfig
are not allowed).
- Add an agent that simulates a GPU machine to the default docker
compose
- Add some makefile utilities like only launching the control plane with
`make docker-compose-mount-up-controlplane` and cleaning up redis DB
dump files.
- Rename datastore client methods to be more datastore-centric rather
than application centric (e.g. `token_exists` instead of
`validate_token`)