Fix preview deployment (#24)
Changing from on push to on pull_request has introduced some bug. This
PR makes sure that the preview is using the most recent commit for the
PR branch. It also improves the EC2 instance name to be the branch
instead of `23/merge`, which is cosmetic as the previous approach
functioned fine.