Add VM Agent metric collectors for other metric streams (#25)
Add data collection tools (samplers and CLI) and VM Agent actors that
collect the metrics and send them to the control plane. Added fake
metric generation ability

Hooking up those actors, creating the backend tables/APIs, and having
them send data to the control plane is not part of this PR.

Added a utility to sync code to a remote machines to make GPU testing
easier. Added a utility to run GPU workloads so we can see utilization
vary.